Lib centric setup

+++ crossBoundaryJNI.sh
@@ -5,26 +5,34 @@
 #
 ############
 
-#add build Java db portion
-#echo "Building Java database for: $1"
+#add build cpp db portion
+#echo "Building cpp database for: $1"
 #./buildALL.sh
 
 #change to app name?
-echo "Running Java analysis on: $1"
-./runJavaPortion.sh $1
-
-cp flows/${1}-flow-summary.csv $2
-
-#run cpp build portion
-echo "Building cpp database for: $1"
-./buildcppDB.sh $1
-
-#run cpp analyze portion
 echo "Running cpp analysis on: $1"
 ./runCPPPortion.sh $1
 
-#todo think of a better way?
-#currently to prevent collisions on this files usage
-rm ${2}/${1}-flow-summary.csv
+if [ -f flows/${1}-flow-summary.csv ]; then
+    #todo change all paths
+    cp flows/${1}-flow-summary.csv $2
 
-#todo find out if these should be combined for a nice overall source (java) to sink (cpp) view?
+    #run cpp build portion
+    echo "Building Java database for: $1"
+    ./buildJavaDB.sh $1
+
+    #run cpp analyze portion
+    echo "Running java analysis on: $1"
+    ./runJavaPortion.sh $1
+
+    #todo think of a better way?
+    #currently to prevent collisions on this files usage
+    rm ${2}/${1}-flow-summary.csv
+
+    #todo find out if these should be combined for a nice overall source (java) to sink (cpp) view?
+
+else
+    echo "---------------------------"
+    echo "Could not run cpp analysis."
+    echo "---------------------------"
+fi
